backend/requirements.txt
psycopg2-binary
PyYAML
pandas
sqlalchemy
Flask
Flask-Cors
gunicorn